Replace the Main Events container with a hooks and cookie based implementation
